+++ source/base64.h
@@ -1,4 +1,4 @@
 
-char *bin2base64(char *data, int byte_count);
-char *base642bin(char *data, int *byte_count);
+char *bin2base64(char *data, size_t byte_count);
+char *base642bin(char *data, size_t *byte_count);
